Update README.md
# Face-Detection
In Python, the Haar Cascade classifier, provided by the OpenCV library, is a popular tool for face detection. Haar Cascades are machine learning-based classifiers that work by training on thousands of positive (face) and negative (non-face) images. This classifier then detects features in a face, such as eyes, nose, and mouth, by running through multiple stages or "cascades."

Face detection is a widely used application in fields such as security, user experience, marketing, and more. It involves identifying and locating human faces in digital images or video frames, and it's commonly used as a preliminary step for more advanced facial recognition, facial analysis, or tracking tasks. 

-->Here are some typical ways face detection is used:
1. Security and Surveillance
2. Photo and Video Tagging
3. Healthcare and Psychology

-->Tools and Libraries for Face Detection<--
* OpenCV
* dlib
* TensorFlow and PyTorch